Changed the backend structure to industry standard
