+++ server/src/database.ts
@@ -1,0 +1,19 @@
+import mongoose from "mongoose";
+import dotenv from "dotenv";
+
+dotenv.config();
+const MONGO_URI = process.env.MONGO_URI;
+
+if (!MONGO_URI) {
+  throw new Error("Missing MONGO_URI! Set it in your environment variables.");
+}
+
+export const connectDB = async () => {
+  mongoose
+    .connect(MONGO_URI, {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    } as mongoose.ConnectOptions)
+    .then(() => console.log("Connected to MongoDB"))
+    .catch((err) => console.error("MongoDB connection error:", err));
+};
